chore: refactors and cleanup for the new worker command
