+++ docs/backends/django-q2.md
@@ -78,7 +78,7 @@ Q_CLUSTER = {
 
 PRODUCTION_PROCESSES = {
     "web": {
-        "BACKEND": "django_prodserver.backends.gunicorn.GunicornServer",
+        "BACKEND": "django_prodserver.backends.servers.gunicorn.GunicornServer",
         "ARGS": {"bind": "0.0.0.0:8000", "workers": "4"},
     },
     "worker": {
