+++ src/django_prodserver/backends/waitress.py
@@ -1,24 +1,15 @@
-import waitress.runner
+"""Deprecated; see ``django_prodserver.backends.servers.waitress``."""
 
-from ..utils import wsgi_app_name
-from .base import BaseServerBackend
+import warnings
 
+from .servers.waitress import WaitressServer
 
-class WaitressServer(BaseServerBackend):
-    """
-    WaitressServer Backend.
+warnings.warn(
+    "django_prodserver.backends.waitress is deprecated; import from "
+    "django_prodserver.backends.servers.waitress instead. This module will be "
+    "removed in django-prodserver 4.0.0.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-    Bypass any Django handling of the command and sends all arguments straight
-    to waitress.
-    """
-
-    def start_server(self, *args: str) -> None:
-        """Start the server."""
-        waitress.runner.run(argv=args)
-
-    def prep_server_args(self) -> list[str]:
-        """Prepare the server args."""
-        args = ["waitress"]
-        args.extend(self.args)
-        args.append(wsgi_app_name())
-        return args
+__all__ = ["WaitressServer"]
