--- src/django_prodserver/backends/workers/base.py
@@ -1,3 +1,0 @@
-from ..base import BaseWorkerBackend
-
-__all__ = ["BaseWorkerBackend"]
