+++ tests/backends/test_celery_flower.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 import pytest
 from django.core.exceptions import ImproperlyConfigured
 
-from django_prodserver.backends.workers.celery import CeleryFlower
+from django_prodserver.backends.servers.flower import CeleryFlower
 
 
 class TestCeleryFlowerImportErrors:
@@ -56,14 +56,14 @@ class TestCeleryFlower:
         with patch.dict(sys.modules, {"flower": Mock()}):
             yield
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_init_with_app(self, mock_import_string):
         """Test CeleryFlower initialization with APP config."""
         mock_app = Mock()
         mock_import_string.return_value = mock_app
 
         server_config = {
-            "BACKEND": "django_prodserver.backends.workers.celery.CeleryFlower",
+            "BACKEND": "django_prodserver.backends.servers.flower.CeleryFlower",
             "APP": "myproject.celery.app",
             "ARGS": {"port": "5555", "address": "0.0.0.0"},
         }
@@ -74,7 +74,7 @@ class TestCeleryFlower:
         assert flower.args == ["--port=5555", "--address=0.0.0.0"]
         mock_import_string.assert_called_once_with("myproject.celery.app")
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_init_without_args(self, mock_import_string):
         """Test CeleryFlower initialization without ARGS."""
         mock_app = Mock()
@@ -85,7 +85,7 @@ class TestCeleryFlower:
         assert flower.app == mock_app
         assert flower.args == []
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_start_server(self, mock_import_string):
         """Test start_server dispatches the flower subcommand on the celery app."""
         mock_app = Mock()
@@ -98,7 +98,7 @@ class TestCeleryFlower:
             argv=["flower", "--port=5555", "--address=0.0.0.0"]
         )
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_start_server_no_args(self, mock_import_string):
         """Test start_server with no args."""
         mock_app = Mock()
@@ -109,7 +109,7 @@ class TestCeleryFlower:
 
         mock_app.start.assert_called_once_with(argv=["flower"])
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_full_workflow(self, mock_import_string):
         """Test the complete workflow from initialization to server start."""
         mock_app = Mock()
@@ -120,19 +120,21 @@ class TestCeleryFlower:
 
         mock_app.start.assert_called_once_with(argv=["flower", "--port=5555"])
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_inheritance(self, mock_import_string):
-        """Test that CeleryFlower inherits from CeleryWorker / BaseWorkerBackend."""
-        from django_prodserver.backends.base import BaseWorkerBackend
-        from django_prodserver.backends.workers.celery import CeleryWorker
+        """Test that CeleryFlower is a server backend, not a worker backend."""
+        from django_prodserver.backends.base import (
+            BaseServerBackend,
+            BaseWorkerBackend,
+        )
 
         mock_import_string.return_value = Mock()
         flower = CeleryFlower(APP="myproject.celery.app")
 
-        assert isinstance(flower, CeleryWorker)
-        assert isinstance(flower, BaseWorkerBackend)
+        assert isinstance(flower, BaseServerBackend)
+        assert not isinstance(flower, BaseWorkerBackend)
 
-    @patch("django_prodserver.backends.workers.celery.import_string")
+    @patch("django_prodserver.backends.servers.flower.import_string")
     def test_start_server_exception_propagation(self, mock_import_string):
         """Test that exceptions from app.start are propagated."""
         mock_app = Mock()
